\name{holzinger19}
\alias{holzinger19}

\title{
Psychometric example data set by Holzinger & Swineford (1936)
}
\description{
19 ability measurements from the classic psychometrics data set of Holzinger and Swineford (1939). The original data set contains 24 variables, however, we excluded the last four to obtain an example data set for four factors (see Harman, 1967). We took this dataset from the R-package psychTools (Revelle & Revelle, 2020).
}
\usage{
  data(holzinger19)
}

\references{
Harman, Harry Horace (1967), Modern factor analysis. Chicago, University of Chicago Press.

Holzinger, K. J., & Swineford, F. (1939). A study in factor analysis: The stability of a bi-factor solution. Supplementary Educational Monographs, no. 48. Chicago: University of Chicago, Department of Education.

Revelle, W. (2020). Package ‘psychTools’.
}
