\name{dimEst-internal}
\title{Internal dimEst functions}

\alias{impCov}

\description{Internal dimEst functions.}

\author{Jonas Haslbeck}

\details{

  These are internal functions.

}

\value{

\code{impCov} returns the model-implied covariance matrix from the output of the \code{fa()} function from the \code{psych} package.

}



\keyword{internal}
